fix(Card): fix toggle-in-card navigation (#165)
* fix: handle onClick in Clickable instead of Card if ToggleItem is rendered

If Card renders a Clickable radio toggle, and has tabIndex (onClick prop), it breaks
arrow-navigation between other Toggle radio Cards. In order to fix it, the onClick is moved to the
Clickable component and name prop is provided to Clickable radios of the same group.

* refactor(card): refactor non-flat card style

* fix(clickable): add focus-ring and focus-ring-inset classes to ToggleItem

Display visual feedback on ToggleItem focus

* feat(card): log deprecation warning if Card receives onClick prop

Click events should be handled by Clickable component, not Card. Before implementing a breaking
change in Card, a deprecation warning is logged if Card receives an onClick prop

* test: update tests with the onClick prop changes

* refactor(dead-toggle): rename 'name' prop in DeadToggle to avoid confusion